Fix DCA Collection Query

+++ src/main/java/org/gooru/nucleus/handlers/dataclass/api/processors/repositories/activejdbc/dbhandlers/StudentPerfInAllClasses.java
@@ -90,7 +90,7 @@ public class StudentPerfInAllClasses implements DBHandler {
     	        if(classData.get(AJEntityBaseReports.COURSE_GOORU_OID) != null){
     	        JsonObject classKPI = new JsonObject();
     	        classKPI.put(AJEntityBaseReports.ATTR_CLASS_ID, classData.get(AJEntityBaseReports.CLASS_GOORU_OID).toString());
-    	        classKPI.put(AJEntityBaseReports.ATTR_TIME_SPENT, Integer.valueOf(classData.get(AJEntityBaseReports.ATTR_TIME_SPENT).toString()));
+    	        classKPI.put(AJEntityBaseReports.ATTR_TIME_SPENT, Long.valueOf(classData.get(AJEntityBaseReports.ATTR_TIME_SPENT).toString()));
     	        classKPI.put(AJEntityBaseReports.ATTR_COMPLETED_COUNT, 0);
     	        classKPI.put(AJEntityBaseReports.ATTR_SCORE, 0);
     	        Object classTotalCount = Base.firstCell(AJEntityCourseCollectionCount.GET_CLASS_ASSESSMENT_COUNT,
@@ -125,7 +125,7 @@ public class StudentPerfInAllClasses implements DBHandler {
 	    	if (!classPerfData.isEmpty()) { 
 	    		classPerfData.forEach(classData -> {
 	    			classKPI.put(AJEntityBaseReports.ATTR_CLASS_ID, classData.get(AJEntityBaseReports.CLASS_GOORU_OID).toString());
-        	        classKPI.put(AJEntityBaseReports.ATTR_TIME_SPENT, Integer.valueOf(classData.get(AJEntityBaseReports.ATTR_TIME_SPENT).toString()));        	        
+        	        classKPI.put(AJEntityBaseReports.ATTR_TIME_SPENT, Long.valueOf(classData.get(AJEntityBaseReports.ATTR_TIME_SPENT).toString()));        	        
         	        Object classTotalCount = Base.firstCell(AJEntityCourseCollectionCount.GET_CLASS_ASSESSMENT_COUNT, 
         	        		classData.get(AJEntityBaseReports.COURSE_GOORU_OID).toString());
         	        classKPI.put(AJEntityBaseReports.ATTR_TOTAL_COUNT, classTotalCount != null ? Integer.valueOf(classTotalCount.toString()) : 0);
@@ -133,8 +133,7 @@ public class StudentPerfInAllClasses implements DBHandler {
 	    	});   
 	      	        LazyList<AJEntityBaseReports> studClass =
 		              AJEntityBaseReports.findBySQL(AJEntityBaseReports.GET_DISTINCT_USERS_IN_CLASS, clId);
-		      studClass.forEach(users -> userList.add(users.getString(AJEntityBaseReports.GOORUUID)));
-		      
+		      studClass.forEach(users -> userList.add(users.getString(AJEntityBaseReports.GOORUUID)));		      
 		      
 		      classPerfList = Base.findAll(AJEntityBaseReports.SELECT_ALL_STUDENT_CLASS_COMPLETION_SCORE, clId,
   					listToPostgresArrayString(userList));
